consolidate USDA_CoA modules
