+++ flowsa/methods/flowbyactivitymethods/USDA_CoA_Cropland.yaml
@@ -31,12 +31,13 @@ url:
     # __year__ is used to trigger a replace function to ubstitute in years
     year: __year__
   key_param: key
-url_replace_fxn: !script_function:USDA_CoA_Cropland CoA_Cropland_URL_helper
-call_response_fxn: !script_function:USDA_CoA_Cropland coa_cropland_call
-parse_response_fxn: !script_function:USDA_CoA_Cropland coa_cropland_parse
+url_replace_fxn: !script_function:USDA_CoA CoA_Cropland_URL_helper
+call_response_fxn: !script_function:USDA_CoA coa_call
+parse_response_fxn: !script_function:USDA_CoA coa_cropland_parse
 years:
 - 2012
 - 2017
+- 2022
 sector_levels:
 - CROPS
 - ECONOMICS      # used to get total irrigated cropland
